Finished user use case DAO

+++ src/main/java/entity/CommonUserFactory.java
@@ -17,7 +17,7 @@ public class CommonUserFactory implements UserFactory {
      * @return a new User instance
      */
     @Override
-    public User create(int userID, String userName, ArrayList<Float> location, String mood,
+    public User create(int userID, String userName, ArrayList<Double> location, String mood,
                        int numActivities, int numMeals) {
         return new User(userID, userName, location, mood, numActivities, numMeals);
     }
